requirements.txt
python-telegram-bot[all]
tweepy
python-dotenv
stripe
black
requests
requests_oauthlib
flask
psycopg2-binary
python-lsp-server
pyright
pynacl
qrcode
farcaster
openai
git+https://github.com/mirceapasoi/hub_py.git
grpcio
protobuf
pynacl
blake3
cryptography
pyflakes
autoflake
git+https://github.com/Bibo-Joshi/ptbstats.git@v2.1
Pillow~=9.5.0
PyHyphen==4.0.3
thefuzz~=0.19
click
ptbstats
websockets
dune-client
ijson
